docs-uml/requirements-docs.txt
# =============================================================================
# docs-uml — Dependencias Python para generación de documentación
# =============================================================================
# Instalar en un entorno virtual aislado (.venv-docs), NUNCA en el proyecto.
#
#   python -m venv .venv-docs
#   .venv-docs/Scripts/Activate.ps1   (Windows)
#   source .venv-docs/bin/activate     (Linux/macOS)
#   pip install -r requirements-docs.txt
# =============================================================================

# Sphinx y tema
sphinx>=8.0,<9.0
furo>=2024.1

# Extensiones Sphinx para diagramas
sphinxcontrib-plantuml>=0.30
sphinxcontrib-mermaid>=1.0

# Manejo de configuración y manifiesto
pyyaml>=6.0
jsonschema>=4.20

# Templating para la vista interactiva
jinja2>=3.1

# Utilidades
pathlib2>=2.3; python_version < "3.12"
